refactor: enhance FinanceCategory functionality with filtering, pagination, and resource collection

+++ app/Services/FinanceCategoryService.php
@@ -3,6 +3,7 @@
 namespace App\Services;
 
 use App\Repositories\Contracts\FinanceCategoryContract;
+use Exception;
 use Symfony\Component\HttpKernel\Exception\HttpException;
 
 class FinanceCategoryService
@@ -12,18 +13,18 @@ class FinanceCategoryService
         $this->financeCategoryRepository = $financeCategoryRepository;
     }
 
-    public function getAllFinanceCategories() {
-        return $this->financeCategoryRepository->all();
+    public function getAllFinanceCategories(?array $filters = []) {
+        return $this->financeCategoryRepository->all($filters);
     }
 
-    public function getFinanceCategoryByParam($param) {
-        return $this->financeCategoryRepository->whereEquals("type", $param);
+    public function getAllPaginatedFinanceCategories(?string $pageSize = null, array $filters = []) {
+        return $this->financeCategoryRepository->paginate($pageSize, $filters);
     }
 
     public function getFinanceCategoryById(string $id) {
         try {
             $user = $this->financeCategoryRepository->findOrFail($id);
-        } catch (\Exception $e) {
+        } catch (Exception $e) {
             throw new HttpException(404, $e->getMessage());
         }
 
@@ -35,21 +36,25 @@ class FinanceCategoryService
     }
 
     public function updateFinanceCategory(string $id, array $data) {
-        $financeCategory = $this->getFinanceCategoryById($id);
-
         try {
+            $financeCategory = $this->getFinanceCategoryById($id);
+
             return $this->financeCategoryRepository->update($id, $data) === true ? $financeCategory->fresh() : false;
-        } catch (\Exception $e) {
+        } catch (Exception $e) {
             throw new HttpException(500, $e->getMessage());
         }
     }
 
     public function deleteFinanceCategory(string $id) {
-        $financeCategory = $this->getFinanceCategoryById($id);
-
         try {
+            $financeCategory = $this->getFinanceCategoryById($id);
+
+            if($financeCategory->financeIncome()->exists() || $financeCategory->financeExpense()->exists()) {
+                throw new Exception("This category cannot be deleted because it contains existing transactions.\n To delete this category, you must first delete or recategorize all associated transactions!");
+            }
+
             return $this->financeCategoryRepository->delete($id) === true ? $financeCategory : false;
-        } catch (\Exception $e) {
+        } catch (Exception $e) {
             throw new HttpException(500, $e->getMessage());
         }
     }
